Move Interval from collections to math
Also removes the now empty collections module.
